Initial git commit for mt_server_2.12
This is the first commit for the mt_server component of the eXintegrator
client server system. It is based on version 2.12 of the system and works
with the 2.12 client. For some documentation of the project either look
at http://www.cdb.riken.jp/scb/documentation/ or use google to search for
eXintegrator.

Known issues with this version of the server:

1. This server application needs to be able to access a postgresql database
containing a big bunch of data in a specific format. I initially used the
libpq++ set of libraries since this project has been written in C++. That
was a bad choice as the libpq++ libraries have not been maintained. Although
there are new C++ libraries, I'm of the opinion that it makes more sense to
move everything to the basic C functions since these will always be the most
up to date.

2. I have used a few Qt classes for handling some of the networking and the
creation and handling of threads. Although not too bad, this seems to be
an unnecessary dependancy and it would probably be better to handle this with
more standard C or C++ calls.

3. The client connection function is somewhat buggered up and can hang up under
certain circumstances. If you look at it you'll see that it is a complete mess
and should be rewritten. Naturally it was one of the first things that I wrote,
at a time when I didn't know very much. I should have fixed it a long time ago;
perhaps using git will make me more confident to modify it.

4. There are probably some possibilities for threads to get tangled up with each
other, as sometimes a thread will take a very long time to finish what it's doing.
Still, this doesn't seem to worry other threads, so perhaps it's not thread related.

5. The authentication method is pretty broken. I initially used the crypt function to
hash passwords; however, I couldn't find it on Windows, so I gave up and wrote my
own somewhat silly hashing function. However, user segregation has never been a
serious issue anyway as the system doesn't really keep any secrets (by choice).

6. I don't know anything about security, so there are probably lots of ways in which
running this server constitutes a security risk. For example it would be trivial to
cause an out of memory error on the server just by sending a never-ending message
(or in fact just by specifying a very long length message). That vulnerability should
be easy to fix by just putting in some sort of limit; but that would entail making
a decision which I don't like doing (or setting up a command line switch, or reading
an environment variable or something..).

7. Some of the code is pretty horrible and difficult to understand. Other bits are
perhaps OK. However, I have a feeling that any trained programmer would have a fit
reading the code and possibly run away screaming. Still, what can I say, it works
pretty well, and I work on it sufficiently infrequently that I have to work out what's
going on from reading the code myself. So it can't be that difficult.

8. The unknown issues. Probably lots.