Cleaned up the ConnectionObject::authenticate function
Removed QSocketDevice dependancies from the authenticate function.
This seems to have removed a problem that I experienced if a user
tried to log in without entering a user name and password. It may
also have fixed some bug that resulted in the system crashing on
certain occasions when a user logged in. I'll have to wait and see
on that though.

There are also some smaller code cleanups here and there, but nothing
major.