Added an expand index by gene function and a small bug fix.
1. Added an expand by gene function. This uses the gene probe set
association defined by the ProbeSetSet2 object (raw/probeSetSet2.h)
to expand the current probe set index to include all probe sets from
genes associated with the current probe sets. Note that this function
(in raw/probeSetSet2.cpp) is very ugly due to a discrepancy between
identifiers used in the database (counting from 1) and the ones used
internally (counting from 0). This means that this function has to be
used with a great deal of care.

2. Related to the index discrepancy described above, I made a small change
to the writeIndex function. The function now subtracts 1 for the values that
are set to the internal clientIndex (which contains array offsets rather than
the db identifiers). This is all very embaressing, and something which could
really do with some rewriting and clarification.
This change does not yet seem to cause any problems; however, there may be some
functions that overcame the old bug that may now not work as they should.